+++ lib/bloom_trade_client.rb
@@ -28,12 +28,12 @@ module BloomTradeClient
     has :jwt_callback, classes: Object
   end
 
-  def self.convert(base_currency:, counter_currency:, type:, user_id: nil)
+  def self.convert(base_currency:, counter_currency:, type:, jwt: nil)
     BloomTradeClient::ExchangeRates::Convert.(
       base_currency: base_currency,
       counter_currency: counter_currency,
       type: type,
-      user_id: user_id
+      jwt: jwt
     )
   end
 end